Refactor helper functions to corresponding go files
