+++ pkg/providers/imagefamily/bootstrap/aksbootstrap.go
@@ -461,77 +461,3 @@ func getFeatureState(enabled bool) nbcontractv1.FeatureState {
 
 	return nbcontractv1.FeatureState_FEATURE_STATE_UNSPECIFIED
 }
-
-// Will move to helper function
-// getIdentityConfig returns the identityConfig object based on the identity inputs.
-func getIdentityConfig(servicePrincipalID string, servicePrincipalSecret string, userAssignedIdentityID string) *nbcontractv1.IdentityConfig {
-	identityConfig := nbcontractv1.IdentityConfig{
-		IdentityType:                nbcontractv1.IdentityType_IDENTITY_TYPE_UNSPECIFIED,
-		ServicePrincipalId:          ptr.String(""),
-		ServicePrincipalSecret:      ptr.String(""),
-		AssignedIdentityId:          ptr.String(""),
-		UseManagedIdentityExtension: ptr.String("false"),
-	}
-
-	if userAssignedIdentityID != "" {
-		identityConfig.IdentityType = nbcontractv1.IdentityType_IDENTITY_TYPE_USER_IDENTITY
-		*identityConfig.AssignedIdentityId = userAssignedIdentityID
-		return &identityConfig
-	}
-
-	if (servicePrincipalID != "" || servicePrincipalID == "msi") && (servicePrincipalSecret != "" || servicePrincipalSecret == base64.StdEncoding.EncodeToString([]byte("msi"))) {
-		identityConfig.IdentityType = nbcontractv1.IdentityType_IDENTITY_TYPE_SERVICE_PRINCIPAL
-		*identityConfig.ServicePrincipalId = servicePrincipalID
-		*identityConfig.ServicePrincipalSecret = servicePrincipalSecret
-		return &identityConfig
-	}
-
-	return &identityConfig
-}
-
-// Will move to helper function
-// getLoadBalancerSKI returns the LoadBalancerSku enum based on the input string.
-func getLoadBalancerSKU(sku string) nbcontractv1.LoadBalancerSku {
-	if strings.EqualFold(sku, "Standard") {
-		return nbcontractv1.LoadBalancerSku_LOAD_BALANCER_SKU_STANDARD
-	} else if strings.EqualFold(sku, "Basic") {
-		return nbcontractv1.LoadBalancerSku_LOAD_BALANCER_SKU_BASIC
-	}
-
-	return nbcontractv1.LoadBalancerSku_LOAD_BALANCER_SKU_UNSPECIFIED
-}
-
-// Will move to helper function
-// getNetworkModeType returns the NetworkMode enum based on the input string.
-func getNetworkModeType(networkMode string) nbcontractv1.NetworkModeType {
-	if strings.EqualFold(networkMode, "transparent") {
-		return nbcontractv1.NetworkModeType_NETWORK_MODE_TRANSPARENT
-	} else if strings.EqualFold(networkMode, "l2bridge") {
-		return nbcontractv1.NetworkModeType_NETWORK_MODE_L2BRIDGE
-	}
-
-	return nbcontractv1.NetworkModeType_NETWORK_MODE_UNSPECIFIED
-}
-
-// Will move to helper function
-// getNetworkPluginType returns the NetworkPluginType enum based on the input string.
-func getNetworkPluginType(networkPlugin string) nbcontractv1.NetworkPluginType {
-	if strings.EqualFold(networkPlugin, "azure") {
-		return nbcontractv1.NetworkPluginType_NETWORK_PLUGIN_TYPE_AZURE
-	} else if strings.EqualFold(networkPlugin, "kubenet") {
-		return nbcontractv1.NetworkPluginType_NETWORK_PLUGIN_TYPE_KUBENET
-	}
-
-	return nbcontractv1.NetworkPluginType_NETWORK_PLUGIN_TYPE_NONE
-}
-
-// Will move to helper function
-func getNetworkPolicyType(networkPolicy string) nbcontractv1.NetworkPolicyType {
-	if strings.EqualFold(networkPolicy, "azure") {
-		return nbcontractv1.NetworkPolicyType_NETWORK_POLICY_TYPE_AZURE
-	} else if strings.EqualFold(networkPolicy, "calico") {
-		return nbcontractv1.NetworkPolicyType_NETWORK_POLICY_TYPE_CALICO
-	}
-
-	return nbcontractv1.NetworkPolicyType_NETWORK_POLICY_TYPE_NONE
-}
